# api/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Rin HTTP API",
    "version": "0.7.0",
    "description": "Authoritative preview wire contract for Rin Protocol v2. Bearer authentication is required only when the server is configured with a token; liveness and readiness probes are always unauthenticated."
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:7374"
    }
  ],
  "paths": {
...
    "/v2/diagnostics": {
      "get": {
        "operationId": "diagnostics",
        "x-rin-sdk-profile": "operational",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Diagnostics"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v2/session/create": {
      "post": {
        "operationId": "create_session",
        "requestBody": {
          "$ref": "#/components/requestBodies/CreateSession"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MutationResult"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "507": {
            "$ref": "#/components/responses/InsufficientStorage"
          }
        }
      }
    },
    "/v2/session/observe": {
      "post": {
        "operationId": "observe",
        "requestBody": {
          "$ref": "#/components/requestBodies/Observe"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MutationResult"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "507": {
            "$ref": "#/components/responses/InsufficientStorage"
          }
        }
      }
    },
    "/v2/agent/propose": {
      "post": {
        "operationId": "propose",
        "requestBody": {
          "$ref": "#/components/requestBodies/Propose"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ProposalResult"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "408": {
            "$ref": "#/components/responses/RequestTimeout"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "507": {
            "$ref": "#/components/responses/InsufficientStorage"
          }
        }
      }
    },
    "/v2/jobs/propose": {
      "post": {
        "operationId": "submit_proposal_job",
        "requestBody": {
          "$ref": "#/components/requestBodies/Propose"
        },
        "responses": {
          "202": {
            "$ref": "#/components/responses/ProposalJobSubmission"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v2/jobs/{job_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/JobId"
        }
      ],
      "get": {
        "operationId": "get_proposal_job",
        "responses": {
          "200": {
            "$ref": "#/components/responses/ProposalJob"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "delete": {
        "operationId": "cancel_proposal_job",
        "responses": {
          "200": {
            "$ref": "#/components/responses/ProposalJob"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v2/generation/jobs": {
      "post": {
        "operationId": "submit_generation_job",
        "requestBody": {
          "$ref": "#/components/requestBodies/Generation"
        },
        "responses": {
          "202": {
            "$ref": "#/components/responses/GenerationJobSubmission"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v2/generation/jobs/{job_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/JobId"
        }
      ],
      "get": {
        "operationId": "get_generation_job",
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenerationJob"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "delete": {
        "operationId": "cancel_generation_job",
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenerationJob"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v2/session/activity": {
      "post": {
        "operationId": "set_actor_activity",
        "requestBody": {
          "$ref": "#/components/requestBodies/SetActorActivity"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MutationResult"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "507": {
            "$ref": "#/components/responses/InsufficientStorage"
          }
        }
      }
    },
    "/v2/world/arbitrate": {
      "post": {
        "operationId": "arbitrate",
        "requestBody": {
          "$ref": "#/components/requestBodies/Arbitrate"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ArbitrationResult"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "507": {
            "$ref": "#/components/responses/InsufficientStorage"
          }
        }
      }
    },
    "/v2/session/get": {
      "post": {
        "operationId": "state",
        "requestBody": {
          "$ref": "#/components/requestBodies/Session"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SessionState"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v2/session/stats": {
      "post": {
        "operationId": "session_stats",
        "requestBody": {
          "$ref": "#/components/requestBodies/Session"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SessionStats"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v2/session/archive": {
      "post": {
        "operationId": "archive_session",
        "requestBody": {
          "$ref": "#/components/requestBodies/ArchiveSession"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ArchiveSession"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v2/session/delete": {
      "post": {
        "operationId": "delete_session",
        "requestBody": {
          "$ref": "#/components/requestBodies/DeleteSession"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/DeleteSession"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v2/session/snapshot": {
      "post": {
        "operationId": "snapshot",
        "requestBody": {
          "$ref": "#/components/requestBodies/Session"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Snapshot"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "507": {
            "$ref": "#/components/responses/InsufficientStorage"
          }
        }
      }
    },
    "/v2/session/restore": {
      "post": {
        "operationId": "restore",
        "requestBody": {
          "$ref": "#/components/requestBodies/Restore"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MutationResult"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "507": {
            "$ref": "#/components/responses/InsufficientStorage"
          }
        }
      }
    },
    "/v2/session/export": {
      "post": {
        "operationId": "export_session",
        "x-rin-sdk-profile": "streaming",
        "requestBody": {
          "$ref": "#/components/requestBodies/Session"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SessionTransfer"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v2/session/import": {
      "post": {
        "operationId": "import_session",
        "x-rin-sdk-profile": "streaming",
        "parameters": [
          {
            "$ref": "#/components/parameters/ExpectedGameId"
          },
          {
            "$ref": "#/components/parameters/ExpectedContentId"
          },
          {
            "$ref": "#/components/parameters/ExpectedContentVersion"
          },
          {
            "$ref": "#/components/parameters/ExpectedContentHash"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/SessionTransfer"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MutationResult"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "408": {
            "$ref": "#/components/responses/RequestTimeout"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "507": {
            "$ref": "#/components/responses/InsufficientStorage"
          }
        }
      }
    },
    "/v2/session/timeline": {
      "post": {
        "operationId": "timeline",
        "requestBody": {
          "$ref": "#/components/requestBodies/Timeline"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Timeline"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v2/session/replay": {
      "post": {
        "operationId": "replay",
        "requestBody": {
          "$ref": "#/components/requestBodies/Replay"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Snapshot"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v2/scheduler/due": {
      "post": {
        "operationId": "due_agents",
        "requestBody": {
          "$ref": "#/components/requestBodies/DueAgents"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/DueAgents"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v2/action/report": {
      "post": {
        "operationId": "report_action",
        "requestBody": {
          "$ref": "#/components/requestBodies/ReportAction"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MutationResult"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "507": {
            "$ref": "#/components/responses/InsufficientStorage"
          }
        }
      }
    },
    "/v2/action/report-batch": {
      "post": {
        "operationId": "report_action_batch",
        "requestBody": {
          "$ref": "#/components/requestBodies/BatchActionReport"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MutationResult"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "507": {
            "$ref": "#/components/responses/InsufficientStorage"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Used when the Rin server is configured with an authentication token."
      }
    },
    "parameters": {
      "JobId": {
        "name": "job_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/Identifier"
        }
      },
...
    },
    "requestBodies": {
      "CreateSession": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CreateSessionRequest"
            }
          }
        }
      },
      "Observe": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ObserveRequest"
            }
          }
        }
      },
      "Propose": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProposeRequest"
            }
          }
        }
      },
      "Generation": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GenerationRequest"
            }
          }
        }
      },
      "SetActorActivity": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SetActorActivityRequest"
            }
          }
        }
      },
      "Arbitrate": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArbitrateRequest"
            }
          }
        }
      },
      "Session": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SessionRequest"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "Diagnostics": {
        "description": "Bounded content-free operational diagnostics.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DiagnosticsResponse"
            }
          }
        }
      },
...
      "MutationResult": {
        "description": "Mutation result.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MutationResultResponse"
            }
          }
        }
      },
      "ProposalResult": {
        "description": "Game-authorized proposal.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProposalResultResponse"
            }
          }
        }
      },
      "ProposalJobSubmission": {
        "description": "Proposal job accepted.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProposalJobSubmissionResponse"
            }
          }
        }
      },
      "ProposalJob": {
        "description": "Proposal job state.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProposalJobResponse"
            }
          }
        }
      },
      "GenerationJobSubmission": {
        "description": "Generation job accepted.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GenerationJobSubmissionResponse"
            }
          }
        }
      },
      "GenerationJob": {
        "description": "Generation job state.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GenerationJobResponse"
            }
          }
        }
      },
      "ArbitrationResult": {
        "description": "Arbitration result.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArbitrationResultResponse"
            }
          }
        }
      },
...
      "BadRequest": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "Unauthorized": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "NotFound": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "RequestTimeout": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "Conflict": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "RequestTooLarge": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "UnsupportedMediaType": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "UnprocessableEntity": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "TooManyRequests": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "InsufficientStorage": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "InternalError": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "ServiceUnavailable": {
        "$ref": "#/components/responses/ErrorResponse"
      },
...
    },
    "schemas": {
...
      "Identifier": {
        "type": "string",
        "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,95}$",
        "minLength": 1,
        "maxLength": 96
      },
...
    }
  },
  "security": [
    {},
    {
      "bearerAuth": []
    }
  ]
}